Stop threads within the main loop
Threads used to be stopped by a separate thread that would periodically monitor
the amount of free threads, and slowly stop them, if they were over a certain
threshold.

Now the exact same thing happens, but it's performed within the main loop, so
the extra thread that performs this task is now removed.

This will save allocating a few extra resources.