Fix memory corruption error
I was passing to printf() a string which contained %something, which means
it tried to access more, to get its parameters, causing, with some compiler
flags, a segmentation fault.

Since this doesn't seem to happen with my regular compilation flags, I did
not spot the issue immediately.

Now they are correctly passed to printf so that they can contain any symbol.