Wait 0.2s before testing if weborf is ready
This saves a lot of time.

Basically checking if weborf was awaiting for the socket was done too soon, so it
was always falling back to sleeping 2s to give it extra time.

Now it waits a little bit so there is an actual chance weborf will be ready by the
time the check is performed.

So the 2s sleep is hit more rarely.